Seperated out the coachDetails hook into it's own file - to prevent repetition
