+++ src/app/pick_tale/[coach]/[path]/[choices]/page.tsx
@@ -1,72 +1,21 @@
+// Choices.tsx
 "use client";
-import { useEffect, useState } from "react";
 import { useParams } from "next/navigation";
 import Link from "next/link";
-import coachData from "@/app/data/coachData.json";
-import type {
-	CoachComponents,
-	ComponentDetails,
-} from "@/app/interfaces/CoachDetails";
+import useCoachDetails from "@/app/hooks/useCoachDetails";
 
 const Choices: React.FC = () => {
-	const [details, setDetails] = useState("");
-	const [detailsTitle, setDetailsTitle] = useState("");
-	const [detailsAudio, setDetailsAudio] = useState("");
-	const [detailsImage, setDetailsImage] = useState("");
 	const { choices, coach, path } = useParams();
-
-	useEffect(() => {
-		if (!coach) return;
-
-		const coachDetails = coachData.details.find(
-			(coachURL: CoachComponents) => coachURL.url === coach,
-		);
-
-		if (!coachDetails) {
-			setDetails("Coach not found");
-			setDetailsTitle("");
-			setDetailsAudio("");
-			setDetailsImage("");
-			return;
-		}
-
-		let selectedDetails: ComponentDetails;
-		if (path === "Football") {
-			selectedDetails =
-				choices === "choice_1"
-					? coachDetails.component_4
-					: coachDetails.component_5;
-		} else {
-			selectedDetails =
-				choices === "choice_1"
-					? coachDetails.component_6
-					: coachDetails.component_7;
-		}
-
-		if (!selectedDetails) {
-			setDetails("No details found");
-			setDetailsTitle("");
-			setDetailsAudio("");
-			setDetailsImage("");
-			return;
-		}
-
-		setDetails(selectedDetails.transcript || "");
-		setDetailsTitle(selectedDetails.title || "");
-		setDetailsAudio(selectedDetails.audio_file || "");
-		setDetailsImage(selectedDetails.image_link || "");
-	}, [coach, choices, path]);
+	const { details, detailsTitle, detailsAudio, detailsImage } = useCoachDetails(
+		{ coach, path, choices } as {
+			coach: string;
+			path: string;
+			choices: string;
+		},
+	);
 
 	if (!coach) {
 		return <div>Loading...</div>;
-	}
-
-	const coachDetails = coachData.details.find(
-		(coachURL: CoachComponents) => coachURL.url === coach,
-	);
-
-	if (!coachDetails) {
-		return <div>Coach not found</div>;
 	}
 
 	return (
@@ -82,8 +31,8 @@ const Choices: React.FC = () => {
 			<h2 className="text-2xl">{detailsTitle}</h2>
 			<br />
 			<p>{details}</p>
-			<p>Audio: {detailsAudio}</p>
-			<p>Image: {detailsImage}</p>
+			<p>{detailsAudio}</p>
+			<p>{detailsImage}</p>
 			<br />
 			<Link href={`/pick_tale/${coach}/${path}/${choices}/end`}>End</Link>
 		</div>
