+++ src/app/pick_tale/[coach]/[path]/page.tsx
@@ -1,68 +1,26 @@
+// Path.tsx
 "use client";
-import { useState, useEffect } from "react";
 import { useParams } from "next/navigation";
 import Link from "next/link";
-import coachData from "@/app/data/coachData.json";
-import type { CoachComponents } from "@/app/interfaces/CoachDetails";
+import useCoachDetails from "@/app/hooks/useCoachDetails";
 
 const Path: React.FC = () => {
-	const [details, setDetails] = useState("");
-	const [detailsTitle, setDetailsTitle] = useState("");
-	const [detailsAudio, setDetailsAudio] = useState("");
-	const [detailsImage, setDetailsImage] = useState("");
 	const { coach, path } = useParams();
-
-	useEffect(() => {
-		if (!coach) return;
-
-		const coachDetails = coachData.details.find(
-			(coachURL: CoachComponents) => coachURL.url === coach,
-		);
-
-		if (!coachDetails) {
-			setDetails("Coach not found");
-			setDetailsTitle("");
-			setDetailsAudio("");
-			setDetailsImage("");
-			return;
-		}
-
-		const selectedDetails =
-			path === "Football" ? coachDetails.component_2 : coachDetails.component_3;
-
-		if (!selectedDetails) {
-			setDetails("No details found");
-			setDetailsTitle("");
-			setDetailsAudio("");
-			setDetailsImage("");
-			return;
-		}
-
-		setDetails(selectedDetails.transcript || "");
-		setDetailsTitle(selectedDetails.title || "");
-		setDetailsAudio(selectedDetails.audio_file || "");
-		setDetailsImage(selectedDetails.image_link || "");
-	}, [coach, path]);
+	const { details, detailsTitle, detailsAudio, detailsImage } = useCoachDetails(
+		{ coach, path } as { coach: string; path: string },
+	);
 
 	if (!coach) {
 		return <div>Loading...</div>;
 	}
 
-	const coachDetails = coachData.details.find(
-		(coachURL: CoachComponents) => coachURL.url === coach,
-	);
-
-	if (!coachDetails) {
-		return <div>Coach not found</div>;
-	}
-
 	return (
 		<div className="w-dvw h-dvh grid place-content-center p-5">
-			<h1 className="text-4xl font-bold underline">{coachDetails.name}</h1>
+			<h1 className="text-4xl font-bold underline">{coach}</h1>
 			<h2 className="text-2xl mt-4">{detailsTitle}</h2>
 			<p className="mt-4">{details}</p>
-			<p>Audio: {detailsAudio}</p>
-			<p>Image: {detailsImage}</p>
+			<p>{detailsAudio}</p>
+			<p>{detailsImage}</p>
 			<ul className="mt-4 space-y-2">
 				<li>
 					<Link href={`/pick_tale/${coach}/${path}/choice_1`}>choice 1</Link>
